requirements.txt
scikit-learn==1.3.2
selenium==4.17.2
streamlit==1.28.0
openai==0.27.10
openai-whisper==20231117
pandas==1.5.3
faiss-cpu==1.7.4
torch==2.1.0
transformers==4.37.2
sentence-transformers==2.2.2
pymupdf==1.24.0
